Separated scripts, added more elements of puzzle two
Added new script files for each puzzle and a Global.cs for any global
variables that need to be accessed by multiple puzzles.

Also added colours and a light to the monitor representation of puzzle
two.

Also added Will's object-picking-up script.

TODO: figure out puzzle two, and how to select the active colour on the
monitor.