Locked camera, added keyboard support
The player's up and down camera movement is now limited (via vertical
clamping in the FPSController) to get rid of the bug where the player
can fly by placing a box directly underneath themself. Added WASD
support for puzzle one and two. It gets really janky on my laptop
though. GetKey is probably not the best way to do it.