Fixed colour counts
Sped up movement speed when using WASD because I'm getting annoyed with
how slow it is after doing it 8 thousand times.

Fixed the distance measurements -- was doing the if statements in the
wrong order so it was only calculating when monitorMode == true

Fixed the colour counts. TODO: right now they all start as 3 until you
hover over that light and then it jumps back to 0. Not sure why.