Made puzzle three logic dynamic
Used an array of boxes for each colour instead of hard coding each box.
Deleted a lot of redundant code. Now if we want to add or remove boxes
from that puzzle, we just have to add them into the scene and give them
the right tag.

The represenation on the monitor is also now dynamic based on the number
of boxes in the scene and how many have been placed on their goal.

Boxes also now disappear when placed on a goal.