Testing in MikeTestRealm
added ability to switch beween two light maps "day" and "night"
this will be implemented into "puzzle 1" for the setting the lighting to
the low power setting during the power outage portion of that puzzle.

this will increase performance and allow us to use a great many more
lights and light types in a scene!

VERY EXCITING